@@ -22,25 +22,25 @@
     "deploy": "liara deploy"
   },
   "dependencies": {
-    "axios": "^0.21.1",
+    "axios": "^0.27.2",
     "formik": "^2.2.9",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-ionicons": "^4.2.0",
     "react-loading": "^2.0.3",
-    "react-loading-skeleton": "^2.2.0",
+    "react-loading-skeleton": "^3.1.0",
     "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "4.0.3",
     "react-spinners": "^0.11.0",
     "redux": "^4.1.1",
     "redux-devtools-extension": "^2.13.9",
     "redux-thunk": "^2.3.0",
-    "simplebar-react": "^2.3.5",
-    "styled-components": "^5.3.1",
-    "typescript": "^4.3.5",
+    "simplebar-react": "^2.4.1",
+    "styled-components": "^5.3.5",
+    "typescript": "^4.7.2",
     "web-vitals": "^1.1.2",
-    "yup": "^0.32.9"
+    "yup": "^0.32.11"
   },
   "eslintConfig": {
     "extends": [
@@ -68,8 +68,8 @@
     "@types/node": "^17.0.35",
     "@types/react": "^18.0.9",
     "@types/react-dom": "^18.0.5",
-    "@types/react-router-dom": "^5.1.8",
-    "@types/styled-components": "^5.1.12",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/styled-components": "^5.1.25",
     "@types/yup": "^0.29.14"
   }
 }
