@@ -1,8 +1,8 @@
-import { Dispatch } from 'redux'
-import * as types from '../types'
+import { Dispatch } from "redux"
+import * as types from "../types"
 
-import * as profileService from '../../services/profile'
-import * as userService from '../../services/user'
+import * as profileService from "../../services/profile"
+import * as userService from "../../services/user"
 
 export const getUserProfile =
   (username: string) => async (dispatch: Dispatch) => {
@@ -19,8 +19,8 @@ export const getUserProfile =
     }
   }
 
-export const updateUserProfile =
-  (id: string, data: object) => async (dispatch: Dispatch) => {
+export const updateUserProfile = (id: string, data: object) => {
+  return async (dispatch: Dispatch) => {
     dispatch({ type: types.UPDATE_USER_PROFILE_REQUEST })
 
     try {
@@ -30,9 +30,10 @@ export const updateUserProfile =
         ? dispatch({ type: types.UPDATE_USER_PROFILE_SUCCESS, user: res.user })
         : dispatch({
             type: types.UPDATE_USER_PROFILE_FAILURE,
-            error: res.message
+            error: res.message,
           })
     } catch (error) {
       dispatch({ type: types.UPDATE_USER_PROFILE_FAILURE, error })
     }
   }
+}
