@@ -22,7 +22,6 @@ const reducer = combineReducers({
 })
 
 const initialState = {}
-
 const middleware = [thunk]
 
 const store = createStore(
@@ -30,5 +29,8 @@ const store = createStore(
   initialState,
   composeWithDevTools(applyMiddleware(...middleware))
 )
+
+export type AppState = ReturnType<typeof store.getState>
+export type AppDispatch = typeof store.dispatch
 
 export default store
