@@ -1,22 +1,35 @@
-import * as types from '../types'
+import { IUser } from "../../types/schemas"
+import * as types from "../types"
 
 interface Action {
   type: string
-  user: object
+  user: IUser
   hasUser: boolean
   error?: string
 }
 
-const initialState = { loading: false, user: {}, error: '', hasUser: false }
+type InitialStateType = {
+  user: IUser | null
+  hasUser: boolean
+  error: string | null
+  loading: boolean
+}
+
+const initialState: InitialStateType = {
+  loading: false,
+  user: null,
+  error: "",
+  hasUser: false,
+}
 
 export function authorizeReducer(state = initialState, action: Action) {
   switch (action.type) {
     case types.GET_AUTHORIZE_USER_REQUEST:
-      return { loading: true, user: {} }
+      return { loading: true, user: null }
     case types.GET_AUTHORIZE_USER_SUCCESS:
       return { loading: false, user: action.user, hasUser: true }
     case types.GET_AUTHORIZE_USER_FAILURE:
-      return { loading: false, user: {}, error: action.error, hasUser: false }
+      return { loading: false, user: null, error: action.error, hasUser: false }
     default:
       return state
   }
